--- .circleci/coverage-targets
@@ -1,2 +1,0 @@
-coverage
-coverage/gojay
