--- .circleci/lint-buildtags
@@ -1,2 +1,0 @@
-
---build-tags=gojay
