+++ go.mod
@@ -3,8 +3,6 @@ module go.lsp.dev/jsonrpc2
 go 1.15
 
 require (
-	github.com/francoispqt/gojay v1.2.13
 	github.com/goccy/go-json v0.3.4
-	github.com/google/go-cmp v0.5.4
 	go.lsp.dev/pkg v0.0.0-20210125030640-b6310ac75a91
 )
