--- key_gojay.go
@@ -1,18 +1,0 @@
-// Copyright 2019 The Go Language Server Authors.
-// SPDX-License-Identifier: BSD-3-Clause
-
-// +build gojay
-
-package jsonrpc2
-
-const (
-	keyCode    = "code"
-	keyError   = "error"
-	keyID      = "id"
-	keyJSONRPC = "jsonrpc"
-	keyMessage = "message"
-	keyMethod  = "method"
-	keyParams  = "params"
-	keyResult  = "result"
-	keyData    = "data"
-)
